indentation and moving imagedepviewer to its own component
